added cors var, separated out contact_us & publish
