+++ services/serverless.yaml
@@ -35,6 +35,7 @@ functions:
       - http:
           path: contact-us
           method: post
+          cors: true
           request:
             template:
               application/json: "{\"message\" : \"$input.path('$.message')\",\"contact\" : \"$input.path('$.contact')\"}"
